Changes by omerg

+++ classes.h
@@ -39,7 +39,7 @@ public:
     tree<Elementy> *employees_pointers_by_salary;
     Company *parent_company;
 
-    Company(int value, int company_id) : value(value), id(company_id), employee_count(0), salary(-1),
+    Company(long double value, int company_id) : value(value), id(company_id), employee_count(0), salary(-1),
                                          employees_pointers_by_salary(nullptr), parent_company(nullptr) {}
 
     ~Company() {
